fix(codegen): address review comments on IterArgCarryAnalyzer (#1813)
- Store `program_` by value instead of const reference to remove lifetime risk
- Add bounds check for `return_vars_[k]` in `ComputeAliasClasses`
- Hoist `EvalConstInt`/`EvalConstTripCount` to `orchestration_analysis` to eliminate duplication between analyzer and codegen
